Lab 0

+++ planets.py
@@ -1,12 +1,11 @@
 def weight_on_planets():
    # write your code here
-   weight = int(input("What do you weigh on earth? "))
+   weight = float(input("What do you weigh on earth? "))
 
    mars = weight * 0.38
    jup = weight * 2.34
 
    print("\nOn Mars you would weigh {} pounds.\nOn Jupiter you would weigh {} pounds.".format(mars, jup))
-   
    
 if __name__ == '__main__':
    weight_on_planets()
